Перенесла получение предложений в App
