+++ src/components/app/app.tsx
@@ -1,4 +1,6 @@
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
+import { useEffect } from 'react';
+import { useAppDispatch } from '../../hooks/store-hooks.ts';
 import MainScreen from '../../pages/main-screen/main-screen.tsx';
 import { AppRoutes, AuthorizationStatus } from '../../const.ts';
 import LoginScreen from '../../pages/login-screen/login-screen.tsx';
@@ -8,8 +10,15 @@ import NotFoundScreen from '../../pages/not-found-screen/not-found-screen.tsx';
 import PrivateRoute from '../private-route/private-route.tsx';
 import ScrollToTop from '../scroll-to-top/scroll-to-top.tsx';
 import Layout from '../layout/layout.tsx';
+import { getCards } from '../../store/action.ts';
 
 function App(): JSX.Element {
+  const dispatch = useAppDispatch();
+
+  useEffect(() => {
+    dispatch(getCards());
+  }, [dispatch]);
+
   return (
     <BrowserRouter>
       <ScrollToTop />
